+++ src/app/channel/services/generator.service.ts
@@ -1,8 +1,8 @@
 import {Injectable} from '@angular/core';
 import {IModel} from '../../model/model.module';
-import {ITemplate} from '../../channel/channel.module';
+import {ITemplate} from '../interfaces/template';
+import {IChannel} from '../interfaces/channel';
 import {IGeneratorResult} from '../interfaces/generator-result';
-import {IChannel} from '../../channel/interfaces/channel';
 import {WebSocketService} from '../../services/websocket.service';
 import {WebSocketMessages} from '../../interfaces/websocket-message';